Refactored gson config
